[1] Add web api to retrieve currency rate for AUD/CNY per yahoo [2] Add Helper method to display AUD and CNY [3] Add CNYPrice on model Product

+++ iloire Facturacion/Models/Helper/CurrencyHelper.cs
@@ -8,7 +8,7 @@ namespace CoffeeInvoice.Models.Helper
 {
 	public static class CurrencyHelper
 	{
-		public static string Currency(this HtmlHelper helper, decimal data, string locale = "en-US", bool woCurrency = false)
+		public static string Currency(this HtmlHelper helper, double data, string locale = "en-US", bool woCurrency = false)
 		{
 			var culture = new System.Globalization.CultureInfo(locale);
